requirements.txt
streamlit
python-dotenv
fpdf
google-generativeai
langchain
